base: Fix implicit floating-point conversion in RandUtil
In `base::RandomizeByPercentage`, a `uint64_t` value is multiplied by a
`double`. With modern toolchains and stricter warning configurations
(e.g., with `-Wimplicit-int-float-conversion` enabled as part of
`-Wconversion`), this implicit conversion can be flagged as a
compile-time error.

This issue was discovered during a toolchain update in ChromeOS, which
was triggered by a dependency upgrade. The stricter compiler settings
caused a build failure at this location, blocking the upgrade. The
details of the downstream change can be seen at:
https://crrev.com/c/6651706

This patch resolves the issue by making the type conversion explicit
using `static_cast<double>`. This clarifies the programmer's intent that
the potential loss of precision from the conversion is acceptable and
makes the code more robust against future toolchain updates and stricter
warning policies.

Bug: b:425266540
Test: tryjob
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6700897
Commit-Queue: Di Wu <diwux@google.com>
Cr-Commit-Position: refs/heads/main@{#1482403}

+++ base/rand_util.h
@@ -111,8 +111,11 @@ T RandomizeByPercentage(T value, double percentage) {
   // adjustment may not fit in a `T`. The clamped value described in pseudocode
   // step (2) above will always fit in a `uint64_t`, so do math in `uint64_t`s.
   const uint64_t abs_value = SafeUnsignedAbs(value);
+  // Explicitly cast to double to avoid implicit conversion warnings on stricter
+  // toolchains. The potential precision loss from converting a large uint64_t
+  // is acceptable for this percentage-based randomization.
   const uint64_t max_abs_adjustment =
-      ClampRound<uint64_t>(abs_value * percentage / 100);
+      ClampRound<uint64_t>(static_cast<double>(abs_value) * percentage / 100.0);
   if (!max_abs_adjustment) {
     return value;
   }
